# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nursing Academic Projects Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectLanguage =
  (typeof ProjectLanguage)[keyof typeof ProjectLanguage];

export const ProjectLanguage = {
  ar: "ar",
  en: "en",
} as const;

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  draft: "draft",
  generating: "generating",
  images_pending: "images_pending",
  payment_pending: "payment_pending",
  pending_approval: "pending_approval",
  approved: "approved",
  rejected: "rejected",
  completed: "completed",
} as const;

export interface Project {
  id: string;
  topic: string;
  studentNames: string[];
  supervisorName: string;
  department: string;
  language: ProjectLanguage;
  status: ProjectStatus;
  generatedContent?: string | null;
  selectedImages?: string[] | null;
  receiptUrl?: string | null;
  universityLogoUrl?: string | null;
  facultyLogoUrl?: string | null;
  docxDownloadUrl?: string | null;
  pptxDownloadUrl?: string | null;
  rejectionReason?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateProjectBodyLanguage =
  (typeof CreateProjectBodyLanguage)[keyof typeof CreateProjectBodyLanguage];

export const CreateProjectBodyLanguage = {
  ar: "ar",
  en: "en",
} as const;

export interface CreateProjectBody {
  topic: string;
  studentNames: string[];
  supervisorName: string;
  department: string;
  language: CreateProjectBodyLanguage;
  universityLogoUrl?: string | null;
  facultyLogoUrl?: string | null;
}

export interface SelectImagesBody {
  /**
   * @minItems 1
   * @maxItems 5
   */
  imageUrls: string[];
}

export interface UploadReceiptBody {
  receiptUrl: string;
}

export interface RejectProjectBody {
  reason: string;
}

export interface ProjectList {
  projects: Project[];
  total: number;
}

export type PexelsImageSrc = {
  medium: string;
  large: string;
};

export interface PexelsImage {
  id: number;
  url: string;
  photographer: string;
  src: PexelsImageSrc;
}

export interface PexelsSearchResult {
  photos: PexelsImage[];
  total_results: number;
}

export interface UploadImageBody {
  dataUri: string;
  folder?: string | null;
}

export interface UploadImageResult {
  url: string;
  publicId: string;
}

export interface AdminStats {
  totalProjects: number;
  pendingApproval: number;
  completed: number;
  rejected: number;
  recentProjects: Project[];
}

export type ListProjectsParams = {
  status?: ListProjectsStatus;
  limit?: number;
  offset?: number;
};

export type ListProjectsStatus =
  (typeof ListProjectsStatus)[keyof typeof ListProjectsStatus];

export const ListProjectsStatus = {
  pending_approval: "pending_approval",
  approved: "approved",
  rejected: "rejected",
  completed: "completed",
} as const;

export type SearchPexelsImagesParams = {
  query: string;
  per_page?: number;
};