# lib/api-zod/src/generated/types/createProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nursing Academic Projects Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProjectBodyLanguage } from "./createProjectBodyLanguage";

export interface CreateProjectBody {
  topic: string;
  studentNames: string[];
  supervisorName: string;
  department: string;
  language: CreateProjectBodyLanguage;
  universityLogoUrl?: string | null;
  facultyLogoUrl?: string | null;
}