# lib/api-zod/src/generated/types/pexelsImage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nursing Academic Projects Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { PexelsImageSrc } from "./pexelsImageSrc";

export interface PexelsImage {
  id: number;
  url: string;
  photographer: string;
  src: PexelsImageSrc;
}