# lib/api-zod/src/generated/types/pexelsSearchResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nursing Academic Projects Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { PexelsImage } from "./pexelsImage";

export interface PexelsSearchResult {
  photos: PexelsImage[];
  total_results: number;
}