# lib/api-zod/src/generated/types/rejectProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nursing Academic Projects Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface RejectProjectBody {
  reason: string;
}