# lib/api-zod/src/generated/types/selectImagesBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nursing Academic Projects Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface SelectImagesBody {
  /**
   * @minItems 1
   * @maxItems 5
   */
  imageUrls: string[];
}