# lib/api-zod/src/generated/types/uploadImageResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nursing Academic Projects Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadImageResult {
  url: string;
  publicId: string;
}